Migrate IDs to int64 type (#816)
A helper function "Int64" is also implemented to assist the migration of
unit tests.

Refer #597.

+++ github/apps_marketplace.go
@@ -29,7 +29,7 @@ type MarketplaceService struct {
 type MarketplacePlan struct {
 	URL                 *string   `json:"url,omitempty"`
 	AccountsURL         *string   `json:"accounts_url,omitempty"`
-	ID                  *int      `json:"id,omitempty"`
+	ID                  *int64    `json:"id,omitempty"`
 	Name                *string   `json:"name,omitempty"`
 	Description         *string   `json:"description,omitempty"`
 	MonthlyPriceInCents *int      `json:"monthly_price_in_cents,omitempty"`
@@ -52,7 +52,7 @@ type MarketplacePurchase struct {
 type MarketplacePlanAccount struct {
 	URL                      *string              `json:"url,omitempty"`
 	Type                     *string              `json:"type,omitempty"`
-	ID                       *int                 `json:"id,omitempty"`
+	ID                       *int64               `json:"id,omitempty"`
 	Login                    *string              `json:"login,omitempty"`
 	Email                    *string              `json:"email,omitempty"`
 	OrganizationBillingEmail *string              `json:"organization_billing_email,omitempty"`
@@ -89,7 +89,7 @@ func (s *MarketplaceService) ListPlans(ctx context.Context, opt *ListOptions) ([
 // ListPlanAccountsForPlan lists all GitHub accounts (user or organization) on a specific plan.
 //
 // GitHub API docs: https://developer.github.com/v3/apps/marketplace/#list-all-github-accounts-user-or-organization-on-a-specific-plan
-func (s *MarketplaceService) ListPlanAccountsForPlan(ctx context.Context, planID int, opt *ListOptions) ([]*MarketplacePlanAccount, *Response, error) {
+func (s *MarketplaceService) ListPlanAccountsForPlan(ctx context.Context, planID int64, opt *ListOptions) ([]*MarketplacePlanAccount, *Response, error) {
 	uri := s.marketplaceURI(fmt.Sprintf("plans/%v/accounts", planID))
 	u, err := addOptions(uri, opt)
 	if err != nil {
@@ -116,7 +116,7 @@ func (s *MarketplaceService) ListPlanAccountsForPlan(ctx context.Context, planID
 // ListPlanAccountsForAccount lists all GitHub accounts (user or organization) associated with an account.
 //
 // GitHub API docs: https://developer.github.com/v3/apps/marketplace/#check-if-a-github-account-is-associated-with-any-marketplace-listing
-func (s *MarketplaceService) ListPlanAccountsForAccount(ctx context.Context, accountID int, opt *ListOptions) ([]*MarketplacePlanAccount, *Response, error) {
+func (s *MarketplaceService) ListPlanAccountsForAccount(ctx context.Context, accountID int64, opt *ListOptions) ([]*MarketplacePlanAccount, *Response, error) {
 	uri := s.marketplaceURI(fmt.Sprintf("accounts/%v", accountID))
 	u, err := addOptions(uri, opt)
 	if err != nil {
