+++ github/repos_deployments.go
@@ -15,7 +15,7 @@ import (
 // Deployment represents a deployment in a repo
 type Deployment struct {
 	URL           *string         `json:"url,omitempty"`
-	ID            *int            `json:"id,omitempty"`
+	ID            *int64          `json:"id,omitempty"`
 	SHA           *string         `json:"sha,omitempty"`
 	Ref           *string         `json:"ref,omitempty"`
 	Task          *string         `json:"task,omitempty"`
@@ -91,7 +91,7 @@ func (s *RepositoriesService) ListDeployments(ctx context.Context, owner, repo s
 // GetDeployment returns a single deployment of a repository.
 //
 // GitHub API docs: https://developer.github.com/v3/repos/deployments/#get-a-single-deployment
-func (s *RepositoriesService) GetDeployment(ctx context.Context, owner, repo string, deploymentID int) (*Deployment, *Response, error) {
+func (s *RepositoriesService) GetDeployment(ctx context.Context, owner, repo string, deploymentID int64) (*Deployment, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/deployments/%v", owner, repo, deploymentID)
 
 	req, err := s.client.NewRequest("GET", u, nil)
@@ -138,7 +138,7 @@ func (s *RepositoriesService) CreateDeployment(ctx context.Context, owner, repo 
 // DeploymentStatus represents the status of a
 // particular deployment.
 type DeploymentStatus struct {
-	ID *int `json:"id,omitempty"`
+	ID *int64 `json:"id,omitempty"`
 	// State is the deployment state.
 	// Possible values are: "pending", "success", "failure", "error", "inactive".
 	State         *string    `json:"state,omitempty"`
@@ -164,7 +164,7 @@ type DeploymentStatusRequest struct {
 // ListDeploymentStatuses lists the statuses of a given deployment of a repository.
 //
 // GitHub API docs: https://developer.github.com/v3/repos/deployments/#list-deployment-statuses
-func (s *RepositoriesService) ListDeploymentStatuses(ctx context.Context, owner, repo string, deployment int, opt *ListOptions) ([]*DeploymentStatus, *Response, error) {
+func (s *RepositoriesService) ListDeploymentStatuses(ctx context.Context, owner, repo string, deployment int64, opt *ListOptions) ([]*DeploymentStatus, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/deployments/%v/statuses", owner, repo, deployment)
 	u, err := addOptions(u, opt)
 	if err != nil {
@@ -191,7 +191,7 @@ func (s *RepositoriesService) ListDeploymentStatuses(ctx context.Context, owner,
 // GetDeploymentStatus returns a single deployment status of a repository.
 //
 // GitHub API docs: https://developer.github.com/v3/repos/deployments/#get-a-single-deployment-status
-func (s *RepositoriesService) GetDeploymentStatus(ctx context.Context, owner, repo string, deploymentID, deploymentStatusID int) (*DeploymentStatus, *Response, error) {
+func (s *RepositoriesService) GetDeploymentStatus(ctx context.Context, owner, repo string, deploymentID, deploymentStatusID int64) (*DeploymentStatus, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/deployments/%v/statuses/%v", owner, repo, deploymentID, deploymentStatusID)
 
 	req, err := s.client.NewRequest("GET", u, nil)
@@ -215,7 +215,7 @@ func (s *RepositoriesService) GetDeploymentStatus(ctx context.Context, owner, re
 // CreateDeploymentStatus creates a new status for a deployment.
 //
 // GitHub API docs: https://developer.github.com/v3/repos/deployments/#create-a-deployment-status
-func (s *RepositoriesService) CreateDeploymentStatus(ctx context.Context, owner, repo string, deployment int, request *DeploymentStatusRequest) (*DeploymentStatus, *Response, error) {
+func (s *RepositoriesService) CreateDeploymentStatus(ctx context.Context, owner, repo string, deployment int64, request *DeploymentStatusRequest) (*DeploymentStatus, *Response, error) {
 	u := fmt.Sprintf("repos/%v/%v/deployments/%v/statuses", owner, repo, deployment)
 
 	req, err := s.client.NewRequest("POST", u, request)
